refactor: move tunnel pkg down one level
While there, reduce unnecessary dependency on external packages.
